# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              },
              "examples": {
                "example1": {
                  "value": {
                    "email": "user@example.com",
                    "password": "Password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Invalid credentials",
                    "value": {
                      "success": false,
                      "message": "Invalid credentials",
                      "error": "Unauthorized",
                      "statusCode": 401,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/login"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Login user",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Airport API",
    "description": "API documentation for the Smart Airport application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "cse"
          },
          "lastName": {
            "type": "string",
            "example": "zag"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "roles": {
            "example": [
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "roles",
          "isVerified"
        ]
      },
...
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4="
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
...
    }
  }
}